Refactored code from 1250 lines down to 250
It is still a lot of lines and could be refactored further. Specifically, the checkBoard function could be 5-10 lines. But a lot of the code is for the computer logic to make unbeatable moves. Soon, I will add 3 levels to the game: easy, hard, unbeatable - which should be fairly easy to implement.